v19
Now the enemies can fire, but the player is the one that controls their firing by pressing 'e', for now. It can hit the player too, with explosion!

Also, the collision works now, and the enemies are "destroyed" by shrinking as before, but also by moving them outside the screen with "-width".